% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMultinomialError.R
\name{addMultinomialError}
\alias{addMultinomialError}
\title{Add multinomial error}
\usage{
addMultinomialError(obs, ess, randomval = NULL)
}
\arguments{
\item{obs}{a vector of starting values}

\item{ess}{effective sample size}

\item{randomval}{seed value for random selection, default = NULL}
}
\description{
Randomly sample from vector that is converted to proportions with input effective sample size.
}
