% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createControlObs.R
\name{createControlObs}
\alias{createControlObs}
\title{Create control list for observations (determines error structures, CVs, samples sizes, etc.)}
\usage{
createControlObs(ESSCAA = 100, CtotCV = 0.05)
}
\arguments{
\item{ESSCAA}{effective sample size for multinomial distribution of catch at age (default=100)}

\item{CtotCV}{coefficient of variation for total catch in weight (default=0.05)}
}
\description{
Create control list for observations (determines error structures, CVs, samples sizes, etc.)
}
