% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOM.R
\name{createOM}
\alias{createOM}
\title{Create Operating Model}
\usage{
createOM(years, nages, nindices, nyear1flag, Rflag, Mflag, Wflag, indexflag,
  catcherrorflag, indexerrorflag, processerrorflag = FALSE)
}
\arguments{
\item{years}{vector of years in operating model (OM)}

\item{nages}{number of ages in OM (assumed to start at age 1)}

\item{nindices}{number of indices}

\item{nyear1flag}{N in first year option}

\item{Rflag}{recruitment option}

\item{Mflag}{natural mortality option}

\item{Wflag}{weights at age option}

\item{indexflag}{list of index options}

\item{catcherrorflag}{option for observation errors in catch}

\item{indexerrorflag}{list of options for observation errors in the indices}

\item{processerrorflag}{option for process error in survival equation, default=FALSE}

\item{Fflag}{fishing mortality option}
}
\description{
Define all necessary settings for creating true population values and observations with noise.
}
