% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genObs.R
\name{genObs}
\alias{genObs}
\title{Generate observations from true OM}
\usage{
genObs(om, controlObs)
}
\arguments{
\item{om}{operating model list from fillTrueOM}

\item{controlObs}{control object for creating observed values}
}
\description{
Generate observations from true OM
}
